requirements.txt
google-cloud-pubsub>=2.0.0
google-cloud-firestore>=2.0.0
google-cloud-bigquery>=3.0.0
yfinance>=0.2.0
pandas>=2.0.0
ta>=0.10.0
python-dateutil>=2.8.0pandas
numpy
scipy
statsmodels
google-cloud-bigquery
kafka-python
pika
jupyter
